updated FilterContainer tests
