+++ src/Filter/__tests__/FilterContainer.test.tsx
@@ -1,0 +1,49 @@
+import * as React from "react";
+import { FilterContainer } from "../FilterContainer";
+import {
+  BasicFilterCategoryModel,
+  AdvancedFilterCategoryModel
+} from "../FilterModels";
+import {
+  mockBasicFilterCategories,
+  mockAdvancedFilterCategoriesAll
+} from "../../../stories/Filter/mocks";
+import { shallow } from "enzyme";
+import toJson from "enzyme-to-json";
+import { BasicFilterContainer } from "../BasicFilterContainer";
+
+describe("FilterContainer", () => {
+  const updateBasicFilter = (
+    basicFilterCategories: BasicFilterCategoryModel[]
+  ) => {
+    wrapper.setProps({ basicFilterCategories });
+  };
+
+  const updateAdvancedFilter = (
+    advancedFilterCategories: AdvancedFilterCategoryModel[]
+  ) => {
+    wrapper.setProps({ advancedFilterCategories });
+  };
+
+  const wrapper = shallow(
+    <FilterContainer
+      basicFilterCategories={mockBasicFilterCategories}
+      onUpdateBasicFilter={updateBasicFilter}
+      advancedFilterCategories={mockAdvancedFilterCategoriesAll}
+      onUpdateAdvancedFilter={updateAdvancedFilter}
+    />,
+    { lifecycleExperimental: true }
+  );
+
+  it("expands the advanced filter", () => {
+    expect(toJson(wrapper, { noKey: false, mode: "deep" })).toMatchSnapshot();
+    for (let i = 0; i < 2; i++) {
+      wrapper
+        .findWhere(node => node.type() === BasicFilterContainer)
+        .dive()
+        .findWhere(node => node.type() === "button")
+        .simulate("click");
+      expect(wrapper).toMatchSnapshot();
+    }
+  });
+});
